Complete the savePhoto function.
Using the try catch function, if status===granted,
First, send an access permission request, and 1. Create an asset. 2. Make an album. 3. If the album does not exist, create an album. (The album's name is "Smily-Cam"). 4. Contrary to step 3, if there is an album, use addAssetsToAlbumAsync to put the asset into the album. If the status is not "granted", setHasPermission(false)=> (a message asking the user to grant access is sent to the user).